Fix skipping of characters when transliterating (v1.24)
The application was ignoring characters if text of greater than one character was being transliterated at once. This occurs when the checkbox was not checked, and the user used the button to initiate the transliteration. The problem was that we were increasing a counter one more than we needed to.