Ignore invisible characters when testing equality
When checking if two strings are equal, ignore invisible characters. This is necessary because when typing in Right to left scripts often an invisible character is inserted without the user's knowledge. If we did not ignore these invisible characters, the user would wonder why the text they typed did not get replaced with the alternative.